requirements.txt
aiogram>=3.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
openai>=1.50.0

